# api.yaml
swagger: "2.0"
info:
  description: >
    Rest-in-contract Project is a product to let you embrace [Consumer-driven contracts](https://martinfowler.com/articles/consumerDrivenContracts.html).
    It is REST in nature so that it fits for integrating with all kind of programming languages.
    For more detail about [Project rest-in-contract](http://blog.airic-yu.com/2062/project-rest-in-contracts), you may have a look in our Project rest-in-contract's Homepage for detail introduction.
        
    This rest-in-contract node module is a module for the Local Contract Server which is the core part of the Rest-in-contract project.
  version: "0.1.0"
  title: "rest-in-contract"
  contact:
    email: "airic.yu@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
host: "localhost"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /contracts:
    post:
      tags:
      - "contract"
      summary: "Add an Contract"
      description: ""
      operationId: "addContract"
      consumes:
      - "application/vnd.js.contract"
      parameters:
      - in: body
        name: body
        description: "An Contract object as JS script that needs to be added"
        required: true
        schema:
          type: string
          example: |-
            module.exports = {
              "name": "testing for hello contract",
              "request": {
                "method": "POST",
                "urlPath": "/hello/apple",
                "queryParameters": [{
                  "name": "a",
                  "value": "b"
                }],
                "body": {
                  "a": "b"
                },
                "headers": {}
              },
              "response": {
                "status": 200,
                "body": {
                  "ok"
                },
                "headers": {}
              }
            }
      responses:
        201:
          description: "Contract created"
    get:
      tags:
      - "contract"
      summary: "Get all Contracts"
      description: ""
      operationId: "getAllContracts"
      responses:
        200:
          description: "OK"
          schema:
            type: array
            items:
              type: string
  /contracts/{contractId}:
    get:
      tags:
      - "contract"
      summary: "Get an Contract"
      description: ""
      operationId: "getContract"
      produces:
      - application/json
      parameters:
      - in: path
        name: contractId
        description: "Contract ID"
        required: true
        type: string
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/ContractHal'
          examples:
            application/json+hal:
              _links:
                self:
                  href: "/api/v1/contracts/94923fbd-9092-4a46-ad65-0d8a2e2f551e"
              id: "94923fbd-9092-4a46-ad65-0d8a2e2f551e"
              name: "testing for hello contract"
              contractScript: "module.exports = { \"id\": \"94923fbd-9092-4a46-ad65-0d8a2e2f551e\", \"name\": \"testing for hello contract\", \"request\": { \"method\": \"POST\", \"urlPath\": \"/hello/apple\", \"queryParameters\": [{\"name\":\"a\",\"value\":\"b\"}], \"body\": { \"a\": \"b\" }, \"headers\": {  } }, \"response\": { \"status\": 200, \"body\": \"ok\", \"headers\": {} } }"
            application/vnd.js.contract: |-
              module.exports = {
                "name": "testing for hello contract",
                "request": {
                  "method": "POST",
                  "urlPath": "/hello/apple",
                  "queryParameters": [{
                    "name": "a",
                    "value": "b"
                  }],
                  "body": {
                    "a": "b"
                  },
                  "headers": {}
                },
                "response": {
                  "status": 200,
                  "body": {
                    "ok"
                  },
                  "headers": {}
                }
              }
        404:
          description: "App not found"
    put:
      tags:
      - "contract"
      summary: "Update an existing Contract"
      description: ""
      operationId: "updateContract"
      consumes:
      - "application/vnd.js.contract"
      parameters:
      - in: path
        name: contractId
        description: "Contract ID"
        required: true
        type: string
      - in: body
        name: body
        description: "An Contract object as JS script that needs to be added"
        required: true
        schema:
          type: string
          example: |-
            module.exports = {
              "name": "testing for hello contract",
              "request": {
                "method": "POST",
                "urlPath": "/hello/apple",
                "queryParameters": [{
                  "name": "a",
                  "value": "b"
                }],
                "body": {
                  "a": "b"
                },
                "headers": {}
              },
              "response": {
                "status": 200,
                "body": {
                  "ok"
                },
                "headers": {}
              }
            }
      responses:
        204:
          description: "Update success"
        404:
          description: "Contract not found"
    delete:
      tags:
      - "contract"
      summary: "Delete an Contract"
      description: ""
      operationId: "deleteContract"
      produces:
      - application/json
      parameters:
      - in: path
        name: contractId
        description: "Contract ID"
        required: true
        type: string
      responses:
        204:
          description: "Delete success"
        404:
          description: "Contract not found"
...
definitions:
...
  ContractHal:
    type: object
    required:
    - id
    - name
    - contractScript
    - _links
    properties:
      id:
        type: string
      name:
        type: string
      contractScript:
        type: string
      _links:
        type: object
        properties:
          self:
            type: object
            properties:
              href:
                type: string
...